fixed issue in the sum of single dimension xarrays

+++ pysd/py_backend/functions.py
@@ -1114,7 +1114,7 @@ def log(x, base):
 def sum(data=None, dim=None):
     if not dim:
         if isinstance(data, xr.DataArray):
-            return np.sum(data).values
+            return np.sum(data.astype(np.float64)).values
         else:
             return np.sum(data)
     else:
